+++ CHANGELOG.md
@@ -5,3 +5,5 @@ Develop
 ----------
 
 * Task #22843733: Header @ds
+* Task #22843736: Hero section @ds
+* Task #22843741: Welcome section @ds